requirements.txt
yfinance==0.1.63
praw==7.2.0
tweepy==3.10.0
pandas==1.2.3
numpy==1.20.2
nltk==3.6.1
plotly.express==0.4.1
pystan==2.19.1.1
fbprophet==0.7.1
datetime==4.3
streamlit==0.81.0
